% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color.schemes}
\alias{color.schemes}
\title{Compatible D3 color schemes}
\format{
An object of class \code{character} of length 38.
}
\usage{
color.schemes
}
\description{
Vector of color schemes available. Most/all d3po functions use d3.interpolate<scheme>.
See \url{https://github.com/d3/d3-scale-chromatic/README.md} for more details on scales.
}
\keyword{datasets}
