\name{d3p_data}
\alias{d3p_data}
\title{Pass a tibble or data frame to 'D3Plus' objects}
\description{This function provides a 1:1 translation between tibble-like objects in R and JSON structured data needed by 'D3Plus' to display a treemap, barchart or any other 'D3Plus' method. This function is useless if \code{d3plus()} is not called first.}
