\name{d3p_id}
\alias{d3p_id}
\title{Defines the ids (such as axis) that 'D3Plus' objects require}
\description{This function provides a 1:1 translation between column names i an R tibble-like object and ids that 'D3Plus' needs to draw axis, bars, areas, etc. This function is useless if \code{d3plus()} is not called first.}
