% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plus-package.R
\name{d3plus-exports}
\alias{d3plus-exports}
\alias{\%>\%}
\alias{JS}
\title{d3plus exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the d3plus package to avoid listing the magrittr
as Depends of d3plus.
}
