% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-package.R
\name{d3po-exports}
\alias{d3po-exports}
\alias{\%>\%}
\alias{JS}
\title{d3po exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the d3po package to avoid listing the magrittr
as Depends of d3po.
}
