% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-package.R
\docType{package}
\name{d3po-package}
\alias{d3po-package}
\title{d3po: Fast and Beautiful Interactive Visualization for 'Markdown' and 'Shiny'}
\description{
Apache licensed alternative to 'Highcharter' which provides functions for both fast and beautiful interactive visualization for 'Markdown' and 'Shiny'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/d3po/}
  \item Report bugs at \url{https://github.com/pachadotdev/d3po/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item John Coene
}

Other contributors:
\itemize{
  \item Ariel Alvarado [contributor]
  \item Sylvain Lesage [contributor]
  \item Curran Kelleher [contributor]
  \item Fernando Becerra [contributor]
  \item  Natural Earth [data contributor]
  \item  R Consortium [funder]
}

}
\keyword{internal}
