% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po.R
\name{d3po-shiny}
\alias{d3po-shiny}
\alias{d3poOutput}
\alias{renderd3po}
\title{Shiny bindings for 'd3po'}
\usage{
d3poOutput(outputId, width = "0", height = "0")

renderd3po(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a d3po object}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using d3po within Shiny
applications and interactive Rmd documents.
}
