% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{d3po_template}
\alias{d3po_template}
\title{Create a new d3po templated project}
\usage{
d3po_template(path)
}
\arguments{
\item{path}{The path to create the new project in}
}
\description{
Create a new d3po templated project
}
