% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.to.adjacency.R
\name{df.to.adjacency}
\alias{df.to.adjacency}
\title{Conversion from data.frame to adjacency matrix}
\usage{
df.to.adjacency(
  df,
  source.column = "source",
  target.column = "target",
  value.column = "value"
)
}
\arguments{
\item{df}{data.frame containing edge data}

\item{source.column}{Name of column containing source nodes. Defaults to "source".}

\item{target.column}{Name of column containing target nodes. Defaults to "target".}

\item{value.column}{Name of column containing edge values. Defaults to "value".}
}
\value{
A list with two components:
\item{matrix}{Adjacency matrix}
\item{labels}{Names of nodes, in same order as rows/columns of adjacency matrix}
}
\description{
Creates an adjacency matrix from an edge list data.frame.
}
\details{
Utilizes a script similar to \url{https://observablehq.com/@d3/chord-diagram} adapted to work with r2d3.
}
\examples{
df = data.frame(source = c("a", "a", "b"),
                target = c("b", "c", "c"),
                value = c(1, 2, 3))
                
df.to.adjacency(df)

}
