% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freedom_house_network}
\alias{freedom_house_network}
\title{Trade Network Coloured by Freedom House Status}
\format{
A \code{igraph} object with 190 vertices (nodes) and 316 edges (arcs).
}
\source{
Adapted from the United Nations (trade volumes) and Freedom House (freedom information).
}
\usage{
freedom_house_network
}
\description{
Connections between countries correspond to the strongest arcs based on the products they export.
The network was trimmed until obtaining an average of four arcs per node.
}
\keyword{datasets}
