% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.r
\name{map_ids}
\alias{map_ids}
\title{Extract the IDs from a Map}
\usage{
map_ids(map)
}
\arguments{
\item{map}{A map object}
}
\value{
A tibble containing IDs and names
}
\description{
Extract the IDs from a Map
}
\examples{
map <- map_ids(maps$south_america$continent)
}
