% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{national}
\alias{national}
\title{National Boundaries Map}
\format{
An \code{sf} object with 202 observations and 4 variables.
}
\source{
Derived from \url{https://www.naturalearthdata.com/}.
}
\usage{
national
}
\description{
National boundaries for all countries in the 'Natural Earth' repository.
The topology has been simplified for better performance in web visualizations and reduced file size.
}
\section{Variables}{

\itemize{
\item \code{continent}: Continent name.
\item \code{country}: Country name.
\item \code{country_iso}: ISO 3166-1 alpha-3 country code.
\item \code{geometry}: Simple feature geometry column.
}
}

\examples{
national[national$country_iso == "GBR", ]
}
\keyword{datasets}
