% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_axis_format}
\alias{po_axis_format}
\title{Axis formatters}
\usage{
po_axis_format(d3po, x = NULL, y = NULL)
}
\arguments{
\item{d3po}{A d3po widget}

\item{x, y}{Optional formatter expressions for x and y axes}
}
\description{
Provide JS formatter expressions for axis tick labels. Accepts strings
starting with \code{JS.} or simple formatter names that will be wrapped for
evaluation on the client-side. Examples:
po_axis_format(p, y = "JS(format-as-billion(value,2))")
po_axis_format(p, y = "format-as-percentage(0)")
}
