% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_background}
\alias{po_background}
\title{Background (Deprecated)}
\usage{
po_background(d3po, background = "#fff")
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{background}{background to add (hex code).}
}
\value{
Appends custom background to an 'htmlwidgets' object
}
\description{
This function is deprecated. Use \code{po_theme(background = ...)} instead.
}
