% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_bar}
\alias{po_bar}
\title{Bar}
\usage{
po_bar(d3po, ..., data = NULL, inherit_daes = TRUE, stack = FALSE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Draw a bar chart.
}
\examples{
if (interactive()) {
  dout <- aggregate(pokemon$name, by = list(
    type = pokemon$type_1,
    color = pokemon$color_1
  ), FUN = length)

  colnames(dout)[3] <- "count"

  # Vertical bars
  d3po(dout) \%>\%
    po_bar(daes(x = type, y = count, color = color)) \%>\%
    po_title("Vertical Bars")

  # Horizontal bars (flipped axes)
  d3po(dout) \%>\%
    po_bar(daes(x = count, y = type, color = color)) \%>\%
    po_title("Horizontal Bars")
}
}
