% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_download}
\alias{po_download}
\title{Download}
\usage{
po_download(d3po, show = TRUE)
}
\arguments{
\item{d3po}{A 'd3po' or 'd3proxy' object.}

\item{show}{Logical indicating whether to show (TRUE) or hide (FALSE) the download button.}
}
\value{
Appends download button settings to an 'htmlwidgets' object
}
\description{
Show/hide the download button.
}
