% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_format}
\alias{po_format}
\title{Format}
\usage{
po_format(d3po, ...)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Named formatting expressions (as quosures). Each name should
correspond to an aesthetic (e.g. \code{x}, \code{y}, \code{tooltip}, etc.).}
}
\description{
Precompute formatted label columns from expressions and attach them to the
widget data. Accepts named expressions like \code{x = round(varx, 2)} or
\code{y = format(varY, big.mark = ",")}. The formatted columns are added to
\code{d3po$x$data} with names \verb{__label_<name>} and registered in
\code{d3po$x$formatted_cols} for the renderer to use.
}
