% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_graph}
\alias{po_graph}
\title{Graph}
\usage{
po_graph(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
Appends nodes arguments to a network-specific 'htmlwidgets' object
}
\description{
Draw a graph.
}
\examples{
d3po(pokemon_graph) \%>\%
  po_graph(daes(size = size, color = color, layout = "nicely")) \%>\%
  po_title("Connections between Pokemon types based on Type 1 and 2")

}
