% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_legend}
\alias{po_legend}
\title{Legend}
\usage{
po_legend(d3po, legend)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{legend}{legend to add.}
}
\value{
Appends custom legend to an 'htmlwidgets' object
}
\description{
Add a legend to a chart.
}
