% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_subtitle}
\alias{po_subtitle}
\title{Subtitle}
\usage{
po_subtitle(d3po, subtitle)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{subtitle}{Character string or JS(...) formatter.}
}
\value{
Appends subtitle settings to an 'htmlwidgets' object
}
\description{
Set a subtitle (or subtitle formatter) for a chart. Accepts a character
string or a JS(...) expression passed through to the client. For
treemaps, a JS function can inspect \code{row.mode} which will be one of
'aggregated', 'flat', or 'drilled'.
}
