% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_title}
\alias{po_title}
\title{Title}
\usage{
po_title(d3po, title)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{title}{Title to add.}
}
\description{
Add a title to a chart.
}
