% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_tooltip}
\alias{po_tooltip}
\title{Tooltip}
\usage{
po_tooltip(d3po, template)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{template}{A character string template or formatter expression.}
}
\value{
Appends tooltip settings to an 'htmlwidgets' object
}
\description{
Set a custom tooltip template for a chart. The template can be a literal
string (e.g. \verb{<b>\{name\}</b><br>Value: \{value\}}) which will be evaluated
server-side by substituting column values.
}
