% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon_graph}
\alias{pokemon_graph}
\title{pokemon_graph}
\format{
A \code{igraph} object with 17 vertices (nodes) and 26 edges (arcs).
}
\source{
Adapted from \code{highcharter} package.
}
\usage{
pokemon_graph
}
\description{
Connections between Pokemon types based on Type 1 and 2.
}
\keyword{datasets}
