% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{Unemployment data}
\format{
An object of class \code{data.frame} with 51 rows and 3 columns.
}
\usage{
unemployment
}
\description{
Unemployment rate by state, July 2019. Data: Bureau of Labor Statistics.
}
\keyword{datasets}
