% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-package.R, R/methods.R
\docType{package}
\name{d3po}
\alias{d3po}
\title{An \code{htmlwidget} interface to the
d3po javascript chart library}
\usage{
d3po(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{d3po need explicit specified data objects formatted as JSON, and this parameter passed it from R.}

\item{...}{Aesthetics to pass, see \code{\link[=daes]{daes()}}}

\item{width}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{Same as width parameter.}

\item{elementId}{Dummy string parameter. Useful when you have two or more charts on the same page.}
}
\value{
Creates a basic 'htmlwidget' object for simple visualization
}
\description{
This function provides 'd3po' methods from R console
}
\author{
Mauricio Vargas
}
