% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-package.R, R/d3po.R
\docType{package}
\name{d3po}
\alias{d3po}
\alias{d3po-package}
\title{An `htmlwidget` interface to the
d3po javascript chart library}
\usage{
d3po(data, width = "100\%", height = "100\%", elementId = NULL)
}
\arguments{
\item{data}{d3po need explicit specified data objects formatted as JSON, and this parameter passed it from R.}

\item{width}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{height}{Same as width parameter.}

\item{elementId}{Dummy string parameter. Useful when you have two or more charts on the same page.}
}
\description{
This function provides 'd3po' methods from R console
}
\examples{
dta <- data.frame(
  id = c("alpha", "alpha", "alpha", "beta", "beta", "beta"),
  x = c(4, 5, 6, 4, 5, 6),
  y = c(7, 25, 13, 7, 8, 13)
)

d3po() \%>\% d3po_data(dta) \%>\% d3po_type("bar")
d3po() \%>\% d3po_data(dta) \%>\% d3po_type("hbar")
d3po() \%>\% d3po_data(dta) \%>\% d3po_type("area")
}
\author{
Mauricio Vargas
}
