% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{daes}
\alias{daes}
\title{Aesthetics}
\usage{
daes(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aspects must be named.}
}
\value{
Aesthetics for the plots such as axis (x,y), group, color and/or size
}
\description{
Aesthetics of the chart.
}
\section{Aesthetics}{

Valid aesthetics (depending on the geom)
\itemize{
\item \code{x}, \code{y}: cartesian coordinates.
\item \code{group}: grouping data.
\item \code{name}: name data.
\item \code{color}: color of geom.
\item \code{size}: size of geom.
\item \code{layout}: layout of geom (nicely, fr, kk, graphopt, drl, lgl, mds, sugiyama), in quotes.
}
}

