% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maps}
\alias{maps}
\title{maps}
\format{
A \code{list} object with 7 elements. The Americas are separated in North America and South America.
}
\source{
Adapted from Natural Earth.
}
\usage{
maps
}
\description{
World, continent and country maps.
}
\keyword{datasets}
