% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_area}
\alias{po_area}
\title{Area}
\usage{
po_area(d3po, ..., data = NULL, inherit_daes = TRUE, stack = FALSE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}

\item{stack}{Whether to stack the series.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot an area chart.
}
\examples{
library(dplyr)

dout <- freedomhouse::country_rating_statuses \%>\%
  group_by(year, status, color) \%>\%
  count()

d3po(dout) \%>\%
  po_area(
    daes(x = year, y = n, group = status, color = color),
    stack = TRUE
  ) \%>\%
  po_title("Evolution of Country Status in Time")
}
