% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_area}
\alias{po_area}
\title{Area}
\usage{
po_area(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot an area chart.
}
\examples{
if (interactive()) {
  dout <- pokemon[pokemon$name \%in\% c(
    "Squirtle", "Wartortle", "Blastoise",
    "Charmander", "Charmeleon", "Charizard",
    "Pikachu", "Raichu"
  ), c("height", "weight", "type_1", "color_1")]

  colnames(dout) <- c("height", "weight", "type", "color")

  d3po(dout) \%>\%
    po_area(daes(
      x = height, y = weight, group = type, color = color
    )) \%>\%
    po_labels(title = "Pokemon Evolution: Weight vs Height by Type")
}
}
