% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_box}
\alias{po_box}
\title{Boxplot}
\usage{
po_box(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Draw a boxplot.
}
\examples{
d3po(pokemon) \%>\%
  po_box(daes(x = type_1, y = speed, group = name, color = color_1)) \%>\%
  po_title("Distribution of Pokemon Speed")

}
