% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_font}
\alias{po_font}
\title{Font}
\usage{
po_font(d3po, family = "Fira Sans", size = 16, transform = "none")
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{family}{family font to use ("Roboto", "Merriweather", etc.).}

\item{size}{size to use (10, 11, 12, etc. overrides auto-sizing).}

\item{transform}{transformation to use for the title ("lowercase", "uppercase", "capitalize", "none").}
}
\value{
Appends custom font to an 'htmlwidgets' object
}
\description{
Edit the font used in a chart.
}
