% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_font}
\alias{po_font}
\title{Font}
\usage{
po_font(d3po, font)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{font}{font to use ("Roboto", "Merriweather", etc.).}
}
\value{
Appends custom font to an 'htmlwidgets' object
}
\description{
Edit the font used in a chart.
}
