% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_geomap}
\alias{po_geomap}
\title{Geomap}
\usage{
po_geomap(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a geomap using sf spatial objects
}
\examples{
if (interactive()) {
  # Using sf objects (requires 'sf' and 'geojsonio' packages)
  library(sf)
  
  # Example 1: North Carolina counties
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
  nc$random <- runif(nrow(nc), min = 10, max = 100)
  
  d3po(nc) \%>\%
    po_geomap(daes(group = FIPS, size = random, tooltip = NAME)) \%>\%
    po_title("Random Values by County")
  
  # Example 2: With custom colors
  nc$my_color <- sample(c("#69b3a2", "#F85888", "#A890F0"), 
                        nrow(nc), replace = TRUE)
  
  d3po(nc) \%>\%
    po_geomap(daes(group = FIPS, size = random, 
                   tooltip = NAME, color = my_color)) \%>\%
    po_title("Counties with Custom Colors")
  
  # Example 3: Gradient coloring based on values
  d3po(nc) \%>\%
    po_geomap(daes(group = FIPS, size = random, 
                   tooltip = NAME, gradient = TRUE)) \%>\%
    po_title("Counties with Gradient")
}
}
