% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_geomap}
\alias{po_geomap}
\title{Geomap}
\usage{
po_geomap(d3po, ..., data = NULL, map = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{map}{map to use (i.e., any valid list or topojson file such as \code{maps$south_america} or \code{jsonlite::fromJSON("south_america.topojson", simplifyVector = F)})}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a geomap
}
\examples{
if (rlang::is_installed("d3pomaps")) {
 library(dplyr)

 dout <- freedom_house \%>\%
   filter(year == 2023, !is.na(iso2c)) \%>\%
   select(id = iso2c, status, color)

 d3po(dout) \%>\% 
   po_geomap(
     daes(group = id, color = color, size = status, tooltip = status),
     map = d3pomaps::maps$south_america$continent
   )
}
}
