% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_geomap}
\alias{po_geomap}
\title{Geomap}
\usage{
po_geomap(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a geomap using sf spatial objects
}
\examples{
if (interactive()) {
 uk <- d3po::subnational
 uk <- uk[uk$country == "United Kingdom", ]
 uk$latitude <- sf::st_coordinates(sf::st_centroid(uk$geometry))[, 2]
 uk$longitude <- sf::st_coordinates(sf::st_centroid(uk$geometry))[, 1]
 uk <- uk[uk$latitude >= 49 & uk$latitude <= 61 & uk$longitude >= -8 & uk$longitude <= 2, ]
 uk$random <- sample(seq_len(nrow(uk)), size = nrow(uk), replace = TRUE)

 # Example 1:
 d3po(uk, width = 800, height = 600) \%>\%
  po_geomap(daes(group = region, size = random, tooltip = region)) \%>\%
  po_labels(title = "Map of UK Regions with Random Values")

 # Example 2: With custom color column
 uk$my_pal <- sample(tintin::tintin_clrs()[1:3], nrow(uk), replace = TRUE)

 d3po(uk, width = 800, height = 600) \%>\%
  po_geomap(daes(group = region, size = random, tooltip = region, color = my_pal)) \%>\%
  po_labels(title = "UK Regions with Custom Colors")

 # Example 3: With custom color vector
 my_color <- tintin::tintin_clrs()[1:3]

 d3po(uk, width = 800, height = 600) \%>\%
  po_geomap(daes(group = region, size = random, tooltip = region, color = my_color)) \%>\%
  po_labels(title = "UK Regions with Custom Colors")

 # Example 4: Gradient coloring based on values
 d3po(uk, width = 800, height = 600) \%>\%
  po_geomap(daes(group = region, size = random, tooltip = region, gradient = TRUE)) \%>\%
  po_labels(title = "UK Regions with Gradient Coloring Based on Random Values")

 # Example 5: Gradient coloring based on values (custom palette)
 d3po(uk, width = 800, height = 600) \%>\%
  po_geomap(daes(group = region, size = random, tooltip = region, color = my_color, gradient = TRUE)) \%>\%
  po_labels(title = "UK Regions with Gradient Coloring Based on Random Values")
}
}
