% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_geomap}
\alias{po_geomap}
\title{Geomap}
\usage{
po_geomap(d3po, ..., data = NULL, map = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{map}{map to use (i.e., \code{system.file("extdata", "worldmap.topojson", package = "d3po")})

map <- system.file("extdata", "south-america.topojson", package = "d3po")
map <- jsonlite::fromJSON(map, simplifyVector = FALSE)
pokemon_mewtwo <- data.frame(id = "CL", value = 1)

d3po(pokemon_mewtwo) \%>\%
po_geomap(daes(group = id, color = value), map = map) \%>\%
po_title("Mewtwo was found in Isla Nueva (New Island, Chile)")}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a geomap
}
