% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_geomap}
\alias{po_geomap}
\title{Geomap}
\usage{
po_geomap(d3po, ..., data = NULL, map = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{map}{map to use (i.e., any valid list or topojson file such as \code{maps$south_america$chile} or \code{jsonlite::fromJSON("chile.topojson", simplifyVector = F)})}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a geomap
}
\examples{
if (interactive()) {
  # Get region IDs and names from the Chile map
  dout <- map_ids(d3po::maps$south_america$chile)

  # Only Magallanes has Pokemon (Mewtwo lives there!)
  dout$pokemon_count <- ifelse(dout$id == "MA", 1L, 0L)
  dout$color <- ifelse(dout$id == "MA", "#F85888", "#e0e0e0")

  d3po(dout) \%>\%
    po_geomap(
      daes(
        group = id, color = color, size = pokemon_count,
        tooltip = name
      ),
      map = d3po::maps$south_america$chile
    ) \%>\%
    po_title("Pokemon Distribution in Chile")
}
}
