% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_labels}
\alias{po_labels}
\title{Labels}
\usage{
po_labels(d3po, align = "center-middle")
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{align}{Label alignment. Must be one of "left-top", "center-middle", or "right-top".}
}
\value{
Appends custom labels to an 'htmlwidgets' object
}
\description{
Edit labels positioning in a treemap.
}
