% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_labels}
\alias{po_labels}
\title{Labels}
\usage{
po_labels(
  d3po,
  align = "left-top",
  x = NULL,
  y = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{align}{Label alignment for treemaps. Must be one of "left-top", "center-middle", or "right-top".}

\item{x}{Optional x-axis label.}

\item{y}{Optional y-axis label.}

\item{title}{Optional title for the chart.}

\item{subtitle}{Optional subtitle for the chart.}
}
\value{
Appends custom labels to an 'htmlwidgets' object
}
\description{
Edit labels positioning in a treemap.
}
