% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_labels}
\alias{po_labels}
\title{Labels}
\usage{
po_labels(d3po, align, valign)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{align}{horizontal alignment (left, center, right, start, middle, end).}

\item{valign}{vertical alignment (top, middle, botton).}
}
\description{
Edit labels positioning in a chart.
}
