% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_scatter}
\alias{po_scatter}
\title{scatter}
\usage{
po_scatter(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot an scatter chart.
}
\examples{
library(dplyr)

dout <- freedomhouse::country_rating_statuses \%>\%
  filter(
    year \%in\% c(1975, 1985, 1995, 2005, 2015),
    country == "Chile"
  ) \%>\%
  mutate(
    inv_civil_liberties = sqrt(1 / civil_liberties),
    inv_political_rights = sqrt(1 / political_rights)
  )

d3po(dout) \%>\%
  po_scatter(
    daes(
      x = inv_civil_liberties, y = inv_political_rights,
      group = year, color = color
    )
  ) \%>\%
  po_title("Evolution of Chile in Time")
}
