% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_scatter}
\alias{po_scatter}
\title{scatter}
\usage{
po_scatter(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot an scatter chart.
}
\examples{
if (interactive()) {
  # Unweighted scatter plot
  d3po(pokemon) \%>\%
    po_scatter(daes(
      x = height, y = weight, group = name, color = color_1
    )) \%>\%
    po_title("Height vs Weight")

  # Log-transformed scatter plot
  pokemon$log_height <- log10(pokemon$height)
  pokemon$log_weight <- log10(pokemon$weight)

  d3po(pokemon) \%>\%
    po_scatter(daes(
      x = log_height, y = log_weight, group = name, color = color_1
    )) \%>\%
    po_title("Log(Height) vs Log(Weight)")

  # Weighted scatterplot by inverse distance from mean
  mean_weight <- mean(pokemon$weight, na.rm = TRUE)
  mean_height <- mean(pokemon$height, na.rm = TRUE)
  pokemon$distance_from_mean_weight <- abs(pokemon$weight - mean_weight)
  pokemon$distance_from_mean_height <- abs(pokemon$height - mean_height)
  pokemon$avg_distance <- (pokemon$distance_from_mean_weight +
    pokemon$distance_from_mean_height) / 2
  pokemon$inverse_distance_from_mean <- 1 / (pokemon$avg_distance + 0.01)

  d3po(pokemon) \%>\%
    po_scatter(daes(
      x = height, y = weight, size = inverse_distance_from_mean,
      group = name, color = color_1
    )) \%>\%
    po_title("Height vs Weight (Size = 1 / Distance from the Mean)")
}
}
