% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3po-api.R
\name{po_treemap}
\alias{po_treemap}
\title{Treemap}
\usage{
po_treemap(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a treemap
}
\examples{
if (rlang::is_installed("freedomhouse")) {
  library(dplyr)

  dout <- freedomhouse::country_rating_statuses \%>\%
    filter(year == 2022) \%>\%
    group_by(status, color) \%>\%
    count()

  d3po(dout) \%>\%
    po_treemap(
      daes(size = n, group = status, color = color)
    ) \%>\%
    po_title("Count of Countries by Continent and Status")
}
}
