% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{po_treemap}
\alias{po_treemap}
\title{Treemap}
\usage{
po_treemap(d3po, ..., data = NULL, inherit_daes = TRUE)
}
\arguments{
\item{d3po}{Either the output of \code{\link[=d3po]{d3po()}} or \code{\link[=d3po_proxy]{d3po_proxy()}}.}

\item{...}{Aesthetics, see \code{\link[=daes]{daes()}}.}

\item{data}{Any dataset to use for plot, overrides data passed
to \code{\link[=d3po]{d3po()}}.}

\item{inherit_daes}{Whether to inherit aesthetics previous specified.}
}
\value{
an 'htmlwidgets' object with the desired interactive plot
}
\description{
Plot a treemap
}
\examples{
dout <- aggregate(pokemon$name, by = list(type = pokemon$type_1,
 color = pokemon$color_1), FUN = length)

colnames(dout)[3] <- "count"

d3po(dout) \%>\%
  po_treemap(daes(size = count, group = type, color = color, tiling = "squarify")) \%>\%
  po_title("Share of Pokemon by main type")
}
